MemoryOutputStream: Avoid integer overflow which may result in incorrect size after expansion

+++ modules/juce_core/streams/juce_MemoryOutputStream.cpp
@@ -93,7 +93,7 @@ char* MemoryOutputStream::prepareToWrite (size_t numBytes)
     if (blockToUse != nullptr)
     {
         if (storageNeeded >= blockToUse->getSize())
-            blockToUse->ensureSize ((storageNeeded + jmin (storageNeeded / 2, (size_t) (1024 * 1024)) + 32) & ~31u);
+            blockToUse->ensureSize ((storageNeeded + jmin (storageNeeded / 2, (size_t) (1024 * 1024)) + 32) & ~(size_t) 31);
 
         data = static_cast<char*> (blockToUse->getData());
     }
